+++ src/server/src/services/getAllAbilityMetadata.ts
